# domains/api.jmjanzen.com/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/me/{id}": {
            "get": {
                "description": "get a representation of me by mode id",
                "produces": [
                    "application/json"
                ],
                "summary": "Get me",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "which mode",
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/me.Me"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/me.NotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/me.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "me.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "me.Me": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "mode": {
                    "type": "string"
                },
                "name": {
                    "$ref": "#/definitions/me.Name"
                }
            }
        },
        "me.Name": {
            "type": "object",
            "properties": {
                "first": {
                    "type": "string"
                },
                "last": {
                    "type": "string"
                },
                "nick": {
                    "type": "string"
                }
            }
        },
        "me.NotFound": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}